# src/api/src/main/java/com/ioteste/api/model/ControllerState.java
package com.ioteste.api.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

@Document(collection = "controller_state")
public class ControllerState {

    @Id
    private String id;

    private boolean enabled;

    public ControllerState() {
    }

    public ControllerState(String id, boolean enabled) {
        this.id = id;
        this.enabled = enabled;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public boolean isEnabled() {
        return enabled;
    }

    public void setEnabled(boolean enabled) {
        this.enabled = enabled;
    }
}