# src/api/src/main/java/com/ioteste/api/model/TemperatureReading.java
package com.ioteste.api.model;

import org.springframework.data.annotation.Id;
import org.springframework.data.mongodb.core.mapping.Document;

import java.time.Instant;

@Document(collection = "temperature_readings")
public class TemperatureReading {

    @Id
    private String id;

    private String roomId;
    private double tempCelsius;
    private double tempFahrenheit;
    private long sourceTs;
    private Instant receivedAt;

    public TemperatureReading() {
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getRoomId() {
        return roomId;
    }

    public void setRoomId(String roomId) {
        this.roomId = roomId;
    }

    public double getTempCelsius() {
        return tempCelsius;
    }

    public void setTempCelsius(double tempCelsius) {
        this.tempCelsius = tempCelsius;
    }

    public double getTempFahrenheit() {
        return tempFahrenheit;
    }

    public void setTempFahrenheit(double tempFahrenheit) {
        this.tempFahrenheit = tempFahrenheit;
    }

    public long getSourceTs() {
        return sourceTs;
    }

    public void setSourceTs(long sourceTs) {
        this.sourceTs = sourceTs;
    }

    public Instant getReceivedAt() {
        return receivedAt;
    }

    public void setReceivedAt(Instant receivedAt) {
        this.receivedAt = receivedAt;
    }
}